refactor: rm TagBadge directory
