+++ packages/bezier-react/src/components/Badge/Badge.module.scss
