+++ packages/bezier-react/src/components/Badge/Badge.types.ts
@@ -1,7 +1,8 @@
+import { type BezierIcon } from '@channel.io/bezier-icons'
+
 import {
   type AlphaBezierComponentProps,
   type ChildrenProps,
-  type ColorProps,
   type SizeProps,
   type VariantProps,
 } from '~/src/types/props'
@@ -9,24 +10,21 @@ import {
 import {
   type TagBadgeSize,
   type TagBadgeVariant,
-} from '~/src/components/TagBadge'
+} from '~/src/components/TagBadgeCommon'
 
-interface TagOwnProps {
+interface BadgeOwnProps {
   /**
-   * Handler to be called when the delete button is clicked.
-   *
-   * The delete button will be not shown if this prop is not provided.
+   * Icon to be shown on the left side of the badge.
    */
-  onDelete?: React.MouseEventHandler
+  icon?: BezierIcon
 }
 
-interface TagProps extends
+interface BadgeProps extends
   AlphaBezierComponentProps,
   ChildrenProps,
   Omit<React.HTMLAttributes<HTMLDivElement>, 'color'>,
   SizeProps<TagBadgeSize>,
   VariantProps<TagBadgeVariant>,
-  ColorProps,
-  TagOwnProps {}
+  BadgeOwnProps {}
 
-export default TagProps
+export default BadgeProps
