+++ packages/bezier-react/src/components/Tag/Tag.tsx
@@ -21,8 +21,8 @@ import {
   TagBadgeVariant,
   getProperTagBadgeBgColor,
   getProperTagBadgeTypo,
-} from '~/src/components/TagBadge/TagBadgeCommon'
-import commonStyles from '~/src/components/TagBadge/TagBadgeCommon/TagBadge.module.scss'
+} from '~/src/components/TagBadgeCommon'
+import commonStyles from '~/src/components/TagBadgeCommon/TagBadge.module.scss'
 import { Text } from '~/src/components/Text'
 
 import type TagProps from './Tag.types'
