--- packages/bezier-react/src/components/TagBadge/index.ts
@@ -1,24 +1,0 @@
-import { Badge } from './Badge'
-import type { BadgeProps } from './Badge'
-import { Tag } from './Tag'
-import type { TagProps } from './Tag'
-import {
-  BadgeColorPreset,
-  TagBadgeBgColorPreset,
-  TagBadgeSize,
-  TagBadgeVariant,
-} from './TagBadgeCommon'
-
-export type {
-  TagProps,
-  BadgeProps,
-}
-
-export {
-  Tag,
-  Badge,
-  TagBadgeSize,
-  TagBadgeVariant,
-  TagBadgeBgColorPreset,
-  BadgeColorPreset,
-}
