+++ packages/bezier-react/src/components/TagBadgeCommon/TagBadge.module.scss
