+++ packages/bezier-react/src/components/TagBadgeCommon/TagBadgeCommon.ts
