+++ packages/bezier-react/src/index.ts
@@ -73,7 +73,9 @@ export * from '~/src/components/Spinner'
 export * from '~/src/components/Stack'
 export * from '~/src/components/Status'
 export * from '~/src/components/Tabs'
-export * from '~/src/components/TagBadge'
+export * from '~/src/components/Tag'
+export * from '~/src/components/Badge'
+export * from '~/src/components/TagBadgeCommon'
 export * from '~/src/components/Text'
 export * from '~/src/components/Toast'
 export * from '~/src/components/Tooltip'
